back/pyproject.toml
[project]
name = "elio-scaffold"
version = "7.0.1"
description = "Elio Scaffold — Agent App development backend"
requires-python = ">=3.12"

dependencies = [
    # Web server
    "fastapi==0.121.3",
    "uvicorn==0.38.0",
    "pydantic==2.12.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "aiofiles==25.1.0",
    "httpx==0.28.1",
    "requests==2.32.5",

    # LLM & Agent frameworks
    "langchain==1.0.8",
    "langchain-core==1.2.15",
    "langchain-community==0.4.1",
    "langchain-openai==1.0.3",
    "langchain-anthropic==1.3.4",
    "langchain-text-splitters==1.0.0",
    "langchain-litellm==0.3.2",
    "langchain-mcp-adapters>=0.1.13",
    "mcp>=1.8.0",
    "litellm==1.66.3",
    "openai==2.11.0",
    "linkup-sdk==0.12.0",

    # Document processing
    "python-docx==1.1.0",
    "python-pptx==0.6.23",
    "PyMuPDF==1.25.3",
    "pdfminer.six==20251230",
    "pdfplumber==0.11.9",
    "openpyxl>=3.1.5",
    "beautifulsoup4==4.12.2",
    "markdownify==1.1.0",
    "tabulate>=0.9.0",
    "tqdm==4.66.6",

    # Document generation (markdown → DOCX)
    "markdown>=3.7",
    "html2docx>=1.6.0",

    # Data
    "pandas>=2.3.0",
    "numpy>=2.3.0",
    "PyYAML==6.0.3",

    # Search & embeddings
    "faiss-cpu==1.13.2",

    # Azure
    "azure-ai-documentintelligence==1.0.2",
    "azure-ai-inference==1.0.0b9",
    "azure-ai-agents==1.2.0b6",
    "azure-cognitiveservices-speech==1.34.0",
    "azure-core==1.32.0",
    "azure-data-tables==12.7.0",
    "azure-identity==1.19.0",
    "azure-keyvault-secrets==4.8.0",
    "azure-storage-blob==12.19.0",
    "langchain-azure-ai==1.0.3",
    "pymongo==4.15.4",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
